вернул случайное отлколнение, чувак теперь работает круто, добавлена вьюшка чувака
